platform_custom/requirements.txt
# Platform API Requirements

# Core
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation
pydantic>=2.7.0
pydantic-settings>=2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2

# Encryption
cryptography==41.0.7

# Utilities
jinja2==3.1.2
pyyaml==6.0.1
